+++ scripts/active_start.bat
@@ -1,3 +1,3 @@
 set RUST_BACKTRACE=1
-..\..\pi_pt\bin\pi_serv -c ../../gaia_wallet_webserver/plugin/wallet/htdocs/.res/wallet,pi_pt,earn/server,earn/xlsx,earn/utils
+..\..\pi_pt\bin\pi_serv_debug -c ../../gaia_wallet_webserver/plugin/wallet/htdocs/.res/wallet,pi_pt,earn/server,earn/xlsx,earn/utils
 pause;
