+++ src/app/config.ts
@@ -35,7 +35,7 @@ export enum DevMode {
     Rinkeby = 'rinkeby'   // rinkeby测试环境
 }
 // tslint:disable-next-line:variable-name
-export const dev_mode:DevMode = DevMode.Ropsten;
+export const dev_mode:DevMode = DevMode.Prod;
 
 // btc网络
 export const btcNetwork = dev_mode === DevMode.Prod ? 'mainnet' : 'testnet';
