+++ src/app/net/push.ts
@@ -5,7 +5,8 @@ import { setBottomLayerReloginMsg, setMsgHandler } from '../../pi/net/ui/con_mgr
 import { backCall, backList, popNew } from '../../pi/ui/root';
 import { getStore, register } from '../store/memstore';
 import { CMD } from '../utils/constants';
-import { getStaticLanguage, logoutAccount, logoutAccountDel, popNewMessage } from '../utils/tools';
+import { getStaticLanguage, popNewMessage } from '../utils/tools';
+import { logoutAccount, logoutAccountDel } from './login';
 import { getServerCloudBalance } from './pull';
 
 // ===================================================== 导入
