+++ src/app/view/mine/setting/setting.ts
@@ -6,8 +6,9 @@ import { popNew } from '../../../../pi/ui/root';
 import { getLang } from '../../../../pi/util/lang';
 import { Forelet } from '../../../../pi/widget/forelet';
 import { Widget } from '../../../../pi/widget/widget';
+import { logoutAccount, logoutAccountDel } from '../../../net/login';
 import { getStore, register, setStore } from '../../../store/memstore';
-import { hasWallet, logoutAccount, logoutAccountDel, popPswBox } from '../../../utils/tools';
+import { hasWallet, popPswBox } from '../../../utils/tools';
 import { backupMnemonic } from '../../../utils/walletTools';
 // ================================================导出
 // tslint:disable-next-line:no-reserved-keywords
